Determine dominate MPs

MPs that perform worse than comparable MPs across all performance metrics are considered 'dominated' as other options are always preferable.

Dom(MSEobj, ..., PMlist = NULL, Refs = NULL, Yrs = NULL)

Arguments

MSEobj

An object of class MSE

...

Names of Performance Metrics (PMs), or other arguments to TradePlot. First PM is recycled if number of PMs is not even

PMlist

Optional list of PM names. Overrides any supplied in ... above

Refs

An optional named list (matching the PM names) with numeric values to override the default Ref values.

Yrs

An optional named list (matching the PM names) with numeric values to override the default Yrs values.

Value

A named list of length 2 with a character vector of non-dominated MPs in MPs and a data.frame of dominated MPs and the names of the relevant dominated MPs in DomMPs

Details

The Dom function compares the probabilities calculated in the performance metric (PM) functions and determines the MPs that have a lower probability across all PMs compared to other MPs of the same management type (e.g., size limit, TAC, etc). Consequently, it is important that all PM functions are constructed so that higher probabilities = better performance (e.g, PNOF is the probability of NOT overfishing)

Author

A. Hordyk

Examples

if (FALSE) {
MSE <- runMSE(MPs=NA) # run all MPs
Nondom <- Dom(MSE, "P10", "LTY", "PNOF")
# Non-dominated MPs
Nondom$MPs

# Dominated MPs
Nondom$DomMPs

}