Class 'Hist'

An object for storing information generated by the end of the historical simulations

Slots

Data

The Data object at the end of the historical period

OMPars

A numeric data.frame with nsim rows with sampled Stock, Fleet, Obs, and Imp parameters.

AtAge

A named list with arrays of dimensions: c(nsim, maxage+1, nyears+proyears) or c(nsim, maxage+1, nyears, nareas)

Length: Length-at-age for each simulation, age, and year

Weight: Weight-at-age for each simulation, age, and year

Select: Selectivity-at-age for each simulation, age, and year

Retention: Retention-at-age for each simulation, age, and year

Maturity: Maturity-at-age for each simulation, age, and year

N.Mortality: Natural mortality-at-age for each simulation, age, and year

Z.Mortality: Total mortality-at-age for each simulation, age, year and area

F.Mortality: Fishing mortality-at-age for each simulation, age, year and area

Fret.Mortality: Fishing mortality-at-age for retained fish for each simulation, age, year and area

Number: Total numbers by simulation, age, year and area

Biomass: Total biomass by simulation, age, year and area

VBiomass: Vulnerable biomass by simulation, age, year and area

SBiomass: Spawning biomass by simulation, age, year and area

Removals: Removals (biomass) by simulation, age, year and area

Landings: Landings (biomass) by simulation, age, year and area

Discards: Discards (biomass) by simulation, age, year and area

TSdata

A named list with population and fleet dynamics:

Number: Total numbers; array dimensions c(nsim, nyears, nareas)

Biomass: Total biomass; array dimensions c(nsim, nyears, nareas)

VBiomass: Vulnerable biomass; array dimensions c(nsim, nyears, nareas)

SBiomass: Spawning Biomass; array dimensions c(nsim, nyears, nareas)

Removals: Removals (biomass); array dimensions c(nsim, nyears, nareas)

Landings: Landings (biomass); array dimensions c(nsim, nyears, nareas)

Discards: Discards (biomass); array dimensions c(nsim, nyears, nareas)

Find: Historical fishing mortality (scale-free); matrix dimensions c(nsim, nyears)

RecDev: Recruitment deviations (historical and projection); matrix dimensions c(nsim, nyears+proyears+maxage)

SPR: Named list with Equilibrium and Dynamic SPR (both matrices iwth dimensions c(nsim, nyears))

Unfished_Equilibrium: A named list with unfished equilibrium numbers and biomass-at-age

Ref

A named list with biological reference points:

ByYear: A named list with asymptotic reference points (i.e., calculated annually without recruitment deviations) all matrices with dimensions nsim by nyears+proyears:

N0: Asymptotic unfished total number

SN0: Asymptotic unfished spawning number

B0: Asymptotic unfished total biomass

SSB0: Asymptotic unfished spawning biomass

VB0: Asymptotic unfished vulnerable biomass

MSY: Asymptotic MSY

FMSY: Fishing mortality corresponding with asymptotic MSY

SSBMSY: Spawning stock biomass corresponding with asymptotic MSY

BMSY: total biomass corresponding with asymptotic MSY

VBMSY: Vulnerable biomass corresponding with asymptotic MSY

F01: Fishing mortality where the change in yield per recruit is 10% of that at F = 0

Fmax: Fishing mortality that maximizes yield per recruit

F_SPR: Fishing mortality corresponding to spawning potential ratio of 20 - 60% in increments of 5%; array dimensions c(nsim, 9, nyears+proyears)

Fcrash: Fishing mortality corresponding to the recruits-per-spawner at the origin of the stock-recruit relationship

Fmed: Fishing mortality corresponding to the median recruits-per-spawner in the historical period

SPRcrash: SPR corresponding to the recruits-per-spawner at the origin of the stock-recruit relationship

Dynamic_Unfished: A named list with dynamic unfished reference points for each simulation and year:

N0: Unfished total numbers

B0: Unfished total biomass

SN0: Unfished spawning numbers

SSB0: Unfished spawning biomass

VB0: Unfished vulnerable biomass

Rec: Unfished recruitment

ReferencePoints: A data.frame with nsim rows with with biological reference points calculated as an average over age-of-maturity ageM years around the current year (i.e. nyears):

N0: Average unfished numbers

B0: Average unfished biomass

SSB0: Average unfished spawning biomass (used to calculate depletion)

SSN0: Average unfished spawning numbers

VB0: Average unfished vulnerable biomass (used to calculate depletion if cpar$control$D='VB')

MSY: Average maximum sustainable yield (equilibrium)

FMSY: Average fishing mortality corresponding with MSY

SSBMSY: Average spawning stock biomass corresponding with MSY

BMSY: Average total biomass corresponding with MSY

VBMSY: Average vulnerable biomass corresponding with MSY

UMSY: Average exploitation rate corresponding with MSY

FMSY_M: Average FMSY/M ratio

SSBMSY_SSB0: Average ratio of SSBMSY to SSB0

BMSY_B0: Average ratio of BMSY to B0

VBMSY_VB0: Average ratio of VBMSY to VB0

RefY: Maximum yield obtained in forward projections with a fixed F

SampPars

A named list with all sampled Stock, Fleet, Obs, and Imp parameters

OM

The OM object (without cpars)

Misc

A list for additional information

Author

A. Hordyk