What objects of this class are available

Generic class finder

Usage

avail(classy, package = NULL, msg = TRUE)

Arguments

classy

A class of object (character string, e.g. 'Fleet')

package

Optional. Names(s) of the package to search for object of class classy. String Default is all openMSE packages. Always searches the global environment as well.

msg

Print messages?

Details

Finds objects of the specified class in the global environment or the openMSE packages.

See also

Author

T. Carruthers

Examples

avail("OM", msg=FALSE)
#> [1] "testOM"
Stocks <- avail("Stock")
#> ✔ Searching for objects of class  Stock  in package:  MSEtool
Fleets <- avail("Fleet")
#> ✔ Searching for objects of class  Fleet  in package:  MSEtool
MPs <- avail("MP")
#> ✔ Searching for objects of class  MP  in package:  MSEtool