Class 'Fleet'

The component of the operating model that controls fishing dynamics

Slots

Name

Name of the Fleet object. Single value. Character string.

Name

Name of the Fleet object. Single value. Character string.

nyears

The number of years for the historical 'spool-up' simulation. Single value. Positive integer

CurrentYr

The current calendar year (final year) of the historical simulations (eg 2011). Single value. Positive integer

EffYears

Years representing join-points (vertices) of time-varying effort. Vector. Non-negative real numbers

EffLower

Lower bound on relative effort corresponding to EffYears. Vector. Non-negative real numbers

EffUpper

Upper bound on relative effort corresponding to EffYears. Vector. Non-negative real numbers

Esd

Additional inter-annual variability in fishing mortality rate. Uniform distribution lower and upper bounds. Non-negative real numbers

qinc

Average percentage change in fishing efficiency (applicable only to forward projection and input controls). Uniform distribution lower and upper bounds. Non-negative real numbers

qcv

Inter-annual variability in fishing efficiency (applicable only to forward projection and input controls). Uniform distribution lower and upper bounds. Non-negative real numbers

L5

Shortest length corresponding to 5 percent vulnerability. Uniform distribution lower and upper bounds. Positive real numbers

LFS

Shortest length that is fully vulnerable to fishing. Uniform distribution lower and upper bounds. Positive real numbers

Vmaxlen

The vulnerability of fish at Stock@Linf. Uniform distribution lower and upper bounds. Fraction

isRel

Selectivity parameters in units of size-of-maturity (or absolute eg cm). Single value. Boolean

LR5

Shortest length corresponding at 5 percent retention. Uniform distribution lower and upper bounds. Non-negative real numbers

LFR

Shortest length that is fully retained. Uniform distribution lower and upper bounds. Non-negative real numbers

Rmaxlen

The retention of fish at Stock@Linf. Uniform distribution lower and upper bounds. Non-negative real numbers

DR

Discard rate - the fraction of caught fish that are discarded. Uniform distribution lower and upper bounds. Fraction

Spat_targ

Distribution of fishing in relation to spatial biomass: fishing distribution is proportional to B^Spat_targ. Uniform distribution lower and upper bounds. Real numbers

MPA

Is Area 1 currently closed (TRUE) or open (FALSE)? Defaults to FALSE

Misc

Miscellaneous list for bio-economic parameters

Creating Object

Objects can be created by calls of the form new('Fleet')

Author

T. Carruthers and A. Hordyk

Examples

showClass('Fleet')

#> Class "Fleet" [package "MSEtool"]
#> 
#> Slots:
#>                                                                             
#> Name:       Name    nyears CurrentYr  EffYears  EffLower  EffUpper       Esd
#> Class: character   numeric   numeric   numeric   numeric   numeric   numeric
#>                                                                             
#> Name:       qinc       qcv        L5       LFS   Vmaxlen     isRel       LR5
#> Class:   numeric   numeric   numeric   numeric   numeric  char.log   numeric
#>                                                                   
#> Name:        LFR   Rmaxlen        DR Spat_targ       MPA      Misc
#> Class:   numeric   numeric   numeric   numeric  char.log      list
#> 
#> Known Subclasses: "OM"