Turing Test

Plots the available data in the Data object together with 5 samples of historical data from the Operating Model (OM) in a random order. The test is used to determine if the data generated by the OM is similar to the fishery data in the Data object. In a well specified OM the user should not be able to visually identify which of the 6 plots is the real fishery data and which are generated by the OM.'

Usage

Turing(OM, Data, wait = TRUE)

TuringMOM(multiHist, Data, wait = TRUE)

Arguments

OM

An object of class OM or class multiHist

Data

An object of class Data or a nested list of Data objects for each stock and fleet

wait

Logical. Wait for key press before next plot?

multiHist

An object of class multiHist. The output of SimulateMOM

Details

In its current form the Turing function does not interpolate missing data in the Data object. Therefore if there are years with missing data, say in the catch time-series, it will be obvious which are the real data and which have been generated by the model. Future versions of the function may include methods to impute missing data for plotting purposes.

The question to ask when examining the plots produced by Turing: do the plots of the 6 data samples look like they are all samples from the same underlying distribution?

Functions

TuringMOM(): Turing function for multi-stock, multi-fleet MOMs

Note

The Turing function was suggested by Andre Punt in his review of one of our recent projects. It is named after the Turing test, developed by Alan Turing in 1950, which is designed to see if a human can detect the difference between human and machine generated information.

Examples

if (FALSE) { # \dontrun{
Turing(MSEtool::testOM, MSEtool::SimulatedData, wait=FALSE)
} # }